refactor: StudioCompass 에서 피커·검색을 갈라낸다 (3,153→2,675줄)
분할 2탄. 앵커드 피커(placePicker/InlinePicker)와 상단 노드 검색
(NodeSearch) 451줄을 StudioPicker.tsx 로, 둘이 본체와 공유하던 보드
기하(BOARD/clampX/clampY)와 종류 글리프(KindGlyph)는 각자 작은 공유
모듈로. 본체에서 받는 것은 타입뿐이라 순환 없음. 시작점 3,523줄 대비
-24%. 동작 변화 0 — 공방 단위 274 · 배치 6 · e2e 13 · 계약 1,639 초록.
